Config state first run fixes
This fixes a problem where when you use this formula's main state `vsftpd` in `init.sls` where the `vsftpd_config` state will fail on the first run due to the path not being available.  I also removed the include of the main `vsftpd` state since the config is broken out in its own state so you can choose to manage only the config while not managing the package and service.